Foram feitas as modificações nas rotas cadastrar produto e atualizar produto. Também foi adicionada as partes dos membros da equipe refatorado o código

+++ src/controladores/produto.js
@@ -4,10 +4,20 @@ const {
   adicionarDados,
   editarDados,
   excluirDados,
+  filtrarDados,
 } = require('../repositorios/consulta')
+const { uploadImagem, excluirImagem } = require('../config/aws_sdk')
+const { extrairBaseUrl } = require('../utils/funcoes')
 
 const listarProdutos = async (req, res) => {
+  const { categoria_id } = req.query
+
   try {
+    if (categoria_id) {
+      const produtos = await filtrarDados('produtos', { categoria_id })
+      return res.status(200).json({ produtos })
+    }
+
     const produtos = await listarDados('produtos')
     return res.status(200).json({ produtos })
   } catch (error) {
@@ -17,6 +27,7 @@ const listarProdutos = async (req, res) => {
 
 const cadastrarProduto = async (req, res) => {
   const produto = req.body
+  const { originalname, mimetype, buffer } = req.file || {}
 
   try {
     const categoriaEncontrada = await buscarDados('categorias', {
@@ -31,7 +42,29 @@ const cadastrarProduto = async (req, res) => {
 
     const [produtoCadastrado] = await adicionarDados('produtos', produto)
 
-    return res.status(201).json(produtoCadastrado)
+    if (!req.file) {
+      return res.status(201).json(produtoCadastrado)
+    }
+
+    const { id } = produtoCadastrado
+    const imagem = await uploadImagem(
+      `produtos/${id}/${originalname}`,
+      buffer,
+      mimetype
+    )
+
+    const baseUrl = `https://f005.backblazeb2.com/file/desafio-final-t04`
+    produto.produto_imagem = `${baseUrl}/${imagem.path}`
+
+    const cadastroCompleto = {
+      tabela: 'produtos',
+      campo: { id },
+      dados: produto,
+    }
+
+    await editarDados(cadastroCompleto)
+
+    return res.status(201).json(cadastroCompleto)
   } catch (error) {
     return res.status(500).json({ mensagem: 'Erro interno do Servidor' })
   }
@@ -40,6 +73,7 @@ const cadastrarProduto = async (req, res) => {
 const editarProduto = async (req, res) => {
   const id = Number(req.params.id)
   const novosDados = req.body
+  const { originalname, mimetype, buffer } = req.file || {}
 
   try {
     const produtoEncontrado = await buscarDados('produtos', { id })
@@ -54,6 +88,26 @@ const editarProduto = async (req, res) => {
 
     if (!categoriaEncontrada) {
       return res.status(404).json({ mensagem: 'Categoria não encontrado.' })
+    }
+
+    const url = produtoEncontrado.produto_imagem
+
+    if (url !== null) {
+      const nomeRef = extrairBaseUrl(url)
+      await excluirImagem(nomeRef)
+      novosDados.produto_imagem = null
+    }
+
+    if (req.file) {
+      const nomeRef = await uploadImagem(
+        `produtos/${id}/${originalname}`,
+        buffer,
+        mimetype
+      )
+
+      const baseUrl = `https://f005.backblazeb2.com/file/desafio-final-t04`
+
+      novosDados.produto_imagem = `${baseUrl}/${nomeRef.path}`
     }
 
     const propriedades = {
@@ -87,6 +141,7 @@ const detalharProduto = async (req, res) => {
 
 const excluirProduto = async (req, res) => {
   const { id } = req.params
+
   try {
     const produtoEncontrado = await buscarDados('produtos', { id })
 
@@ -94,6 +149,23 @@ const excluirProduto = async (req, res) => {
       return res.status(404).json({
         mensagem: 'O produto informado não existe.',
       })
+    }
+
+    const produtoVendido = await buscarDados('pedido_produtos', {
+      produto_id: id,
+    })
+
+    if (produtoVendido) {
+      return res.status(400).json({
+        mensagem: 'Há pedidos em amdamento com este produto',
+      })
+    }
+
+    const url = produtoEncontrado.produto_imagem
+
+    if (url !== null) {
+      const nomeRef = extrairBaseUrl(url)
+      await excluirImagem(nomeRef)
     }
 
     await excluirDados('produtos', { id })
